Switch PQSDKd plugin from server to client mode
In previous design the PQSDK:PQSDKd plugin was working
as a server, which caused number of problems. This patch
changes it and PQSDK:PQSDKd plugin is a client that connects
to PQSDKd.

In order to do that I've added structure called comm_t with
supporting functions. Structure represents communication
context which stores data that are needed to connect a
socket and perform data exchange. There can be multiple
communication sockets (2 by default), which means that there
PQSDK:PQSDKD will open 2 connections to PQSDKd. It is safe
to use both connections in parralel, which makes data exchange
faster in case multiple clients want to use PQSDKD to perform
post-quantum operation.

Also, plugin uses now stream_manager facility implemented in
strongSwan.